+++ src/pages/post/components/PostBottom.tsx
@@ -2,7 +2,8 @@ import CustomDivider from '@shared/ui/CustomDivider';
 import { ToggleButton } from './toggleButton';
 import styled from 'styled-components';
 import IcZoomIn from '@shared/assets/icon/ic-zoom-in.svg?react';
-import { useEffect, useRef } from 'react';
+import IcZoomOut from '@shared/assets/icon/ic-zoom-out.svg?react';
+import { useEffect, useRef, useState } from 'react';
 
 interface PostBottomProps {
   imgZoom: boolean;
@@ -30,6 +31,7 @@ const PostBottom = ({
   };
 
   const textAreaRef = useRef<HTMLTextAreaElement | null>(null);
+  const [, forceUpdate] = useState({});
 
   useEffect(() => {
     if (textAreaRef.current) {
@@ -41,11 +43,13 @@ const PostBottom = ({
 
   const changeTextArea = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
     textStateRef.current = e.target.value;
+    forceUpdate({});
   };
+
   return (
     <>
       <ZoomButton $imgZoom={imgZoom} onClick={handleImgZoom}>
-        <IcZoomStyle />
+        {imgZoom ? <IcZoomOutStyle /> : <IcZoomInStyle />}
       </ZoomButton>
       <TextArea
         ref={textAreaRef}
@@ -58,8 +62,8 @@ const PostBottom = ({
         <ToggleButton buttonState={buttonState} setButtonState={setButtonState} />
         <SaveStyleButton
           onClick={handleClose}
-          disabled={textStateRef === undefined}
-          $textState={!!textStateRef}
+          disabled={textStateRef.current === undefined}
+          $textState={!!textStateRef.current}
         >
           스타일 저장하기
         </SaveStyleButton>
@@ -68,14 +72,7 @@ const PostBottom = ({
   );
 };
 
-const ZoomButton = styled.button<{ $imgZoom: boolean }>`
-  width: 5.924vh;
-  height: 5.924vh;
-  margin-left: 1.5vw;
-  position: absolute;
-  z-index: 10001;
-  top: ${({ $imgZoom }) => ($imgZoom === false ? '46vh' : '50.5vh')};
-`;
+export default PostBottom;
 
 const TextArea = styled.textarea`
   width: 100%;
@@ -90,7 +87,7 @@ const TextArea = styled.textarea`
 
 const BottomDiv = styled.div`
   padding: 0px 5.128vw 0px 5.128vw;
-  margin-top: 1.896vh;
+  margin-top: 3.896vh;
 `;
 const SaveStyleButton = styled.button<{ $textState: boolean | undefined }>`
   width: 100%;
@@ -102,7 +99,14 @@ const SaveStyleButton = styled.button<{ $textState: boolean | undefined }>`
   color: black;
 `;
 
-const IcZoomStyle = styled(IcZoomIn)`
+const IcZoomInStyle = styled(IcZoomIn)`
+  &:hover {
+    path {
+      stroke: ${({ theme }) => theme.colors.green500};
+    }
+  }
+`;
+const IcZoomOutStyle = styled(IcZoomOut)`
   &:hover {
     path {
       stroke: ${({ theme }) => theme.colors.green500};
@@ -110,4 +114,16 @@ const IcZoomStyle = styled(IcZoomIn)`
   }
 `;
 
-export default PostBottom;
+const ZoomButton = styled.button<{ $imgZoom: boolean }>`
+  width: 5.924vh;
+  height: 5.924vh;
+  margin-left: 1.5vw;
+  position: absolute;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 10001;
+  border-radius: 50%;
+  background-color: ${({ theme }) => theme.colors.gray800};
+  top: ${({ $imgZoom }) => ($imgZoom === false ? '46vh' : '50.5vh')};
+`;
