+++ src/shared/ui/CreatePostImageCarousel.tsx
@@ -2,7 +2,7 @@ import styled from 'styled-components';
 import 'react-responsive-carousel/lib/styles/carousel.min.css'; // requires a loader
 import ProgressBar from '@shared/ui/ProgressBar';
 import { Carousel } from 'react-responsive-carousel';
-
+import { useState } from 'react';
 const CreatePostImageCarousel = ({
   images,
   height = '60vh',
